Split table filter functions into multiple files
