+++ src/function/function_list.cpp
@@ -73,12 +73,12 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_SET(InternalDecompressIntegralUintegerFun),
 	DUCKDB_SCALAR_FUNCTION_SET(InternalDecompressIntegralUsmallintFun),
 	DUCKDB_SCALAR_FUNCTION_SET(InternalDecompressStringFun),
-	DUCKDB_SCALAR_FUNCTION(InternalTableFilterBloomFilterFun),
-	DUCKDB_SCALAR_FUNCTION(InternalTableFilterDynamicFun),
-	DUCKDB_SCALAR_FUNCTION(InternalTableFilterOptionalFun),
-	DUCKDB_SCALAR_FUNCTION(InternalTableFilterPerfectHashJoinFun),
-	DUCKDB_SCALAR_FUNCTION(InternalTableFilterPrefixRangeFun),
-	DUCKDB_SCALAR_FUNCTION(InternalTableFilterSelectivityOptionalFun),
+	DUCKDB_SCALAR_FUNCTION(TableFilterBloomFilterFun),
+	DUCKDB_SCALAR_FUNCTION(TableFilterDynamicFun),
+	DUCKDB_SCALAR_FUNCTION(TableFilterOptionalFun),
+	DUCKDB_SCALAR_FUNCTION(TableFilterPerfectHashJoinFun),
+	DUCKDB_SCALAR_FUNCTION(TableFilterPrefixRangeFun),
+	DUCKDB_SCALAR_FUNCTION(TableFilterSelectivityOptionalFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(AddFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(AnyValueFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET_ALIAS(ArbitraryFun),
