+++ src/include/duckdb/planner/filter/selectivity_optional_filter.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "duckdb/planner/filter/optional_filter.hpp"
-#include "duckdb/planner/filter/tablefilter_internal_functions.hpp"
+#include "duckdb/planner/filter/table_filter_functions.hpp"
 
 namespace duckdb {
