+++ src/optimizer/expression_heuristics.cpp
@@ -3,7 +3,7 @@
 
 #include "duckdb/planner/expression/list.hpp"
 #include "duckdb/planner/filter/expression_filter.hpp"
-#include "duckdb/planner/filter/tablefilter_internal_functions.hpp"
+#include "duckdb/planner/filter/table_filter_functions.hpp"
 
 namespace duckdb {
