+++ src/planner/filter/dynamic_filter.cpp
@@ -1,6 +1,6 @@
 #include "duckdb/planner/filter/dynamic_filter.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
-#include "duckdb/planner/filter/tablefilter_internal_functions.hpp"
+#include "duckdb/planner/filter/table_filter_functions.hpp"
 
 namespace duckdb {
