+++ src/planner/filter/optional_filter.cpp
@@ -2,7 +2,7 @@
 #include "duckdb/planner/filter/optional_filter.hpp"
 #include "duckdb/planner/expression.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
-#include "duckdb/planner/filter/tablefilter_internal_functions.hpp"
+#include "duckdb/planner/filter/table_filter_functions.hpp"
 
 namespace duckdb {
