+++ src/planner/filter/prefix_range_filter.cpp
@@ -15,7 +15,7 @@
 #include "duckdb/common/uhugeint.hpp"
 #include "duckdb/main/client_context.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
-#include "duckdb/planner/filter/tablefilter_internal_functions.hpp"
+#include "duckdb/planner/filter/table_filter_functions.hpp"
 #include "duckdb/storage/buffer_manager.hpp"
 
 namespace duckdb {
