+++ src/planner/filter/table_filter_prefix_range_function.cpp
@@ -1,0 +1,114 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/planner/filter/table_filter_prefix_range_function
+//
+//
+//===----------------------------------------------------------------------===//
+
+#include "duckdb/planner/filter/table_filter_functions.hpp"
+#include "table_filter_function_helpers.hpp"
+
+#include "duckdb/execution/expression_executor_state.hpp"
+#include "duckdb/planner/expression/bound_function_expression.hpp"
+#include "duckdb/storage/statistics/numeric_stats.hpp"
+
+namespace duckdb {
+
+PrefixRangeFunctionData::PrefixRangeFunctionData(optional_ptr<PrefixRangeFilter> filter_p,
+                                                 const string &key_column_name_p, const LogicalType &key_type_p,
+                                                 float selectivity_threshold_p, idx_t n_vectors_to_check_p)
+    : filter(filter_p), key_column_name(key_column_name_p), key_type(key_type_p),
+      selectivity_threshold(selectivity_threshold_p), n_vectors_to_check(n_vectors_to_check_p) {
+}
+
+unique_ptr<FunctionData> PrefixRangeFunctionData::Copy() const {
+	return make_uniq<PrefixRangeFunctionData>(filter, key_column_name, key_type, selectivity_threshold,
+	                                          n_vectors_to_check);
+}
+
+bool PrefixRangeFunctionData::Equals(const FunctionData &other_p) const {
+	auto &other = other_p.Cast<PrefixRangeFunctionData>();
+	return filter.get() == other.filter.get() && key_column_name == other.key_column_name && key_type == other.key_type;
+}
+
+idx_t SelectPrefixRange(Vector &input, const PrefixRangeFunctionData &func_data, SelectionVector &result_sel,
+                        idx_t count) {
+	D_ASSERT(func_data.filter);
+	return func_data.filter->LookupKeys(input, result_sel, count);
+}
+
+static unique_ptr<FunctionLocalState>
+PrefixRangeInitLocalState(ExpressionState &state, const BoundFunctionExpression &expr, FunctionData *bind_data) {
+	auto &data = bind_data->Cast<PrefixRangeFunctionData>();
+	if (!data.filter) {
+		return nullptr;
+	}
+	return InitSelectivityTrackingLocalState(data.n_vectors_to_check, data.selectivity_threshold);
+}
+
+static void PrefixRangeFunction(DataChunk &args, ExpressionState &state, Vector &result) {
+	auto &func_expr = state.expr.Cast<BoundFunctionExpression>();
+	auto &func_data = func_expr.bind_info->Cast<PrefixRangeFunctionData>();
+	auto local_state_ptr = ExecuteFunctionState::GetFunctionState(state);
+	auto tracking_state = local_state_ptr ? &local_state_ptr->Cast<SelectivityTrackingLocalState>() : nullptr;
+
+	auto count = args.size();
+
+	if (!func_data.filter || !func_data.filter->IsInitialized()) {
+		SetAllTrue(args, result);
+		return;
+	}
+
+	auto &input = args.data[0];
+	ExecuteWithSelectivityTracking(args, result, tracking_state, [&] {
+		SelectionVector lookup_sel(count);
+		auto found_count = SelectPrefixRange(input, func_data, lookup_sel, count);
+		SelectionToBooleanResult(count, lookup_sel, found_count, result);
+		return found_count;
+	});
+}
+
+ScalarFunction PrefixRangeScalarFun::GetFunction(const LogicalType &input_type) {
+	ScalarFunction func(NAME, {input_type}, LogicalType::BOOLEAN, PrefixRangeFunction, TableFilterFunctions::Bind);
+	func.SetInitStateCallback(PrefixRangeInitLocalState);
+	func.SetNullHandling(FunctionNullHandling::SPECIAL_HANDLING);
+	func.SetFilterPruneCallback(PrefixRangeScalarFun::FilterPrune);
+	func.serialize = TableFilterFunctionSerialize;
+	func.deserialize = TableFilterFunctionDeserialize;
+	return func;
+}
+
+string PrefixRangeScalarFun::ToString(const string &column_name, const string &key_column_name) {
+	return column_name + " IN PRF(" + key_column_name + ")";
+}
+
+FilterPropagateResult PrefixRangeScalarFun::FilterPrune(const FunctionStatisticsPruneInput &input) {
+	if (!input.bind_data) {
+		return FilterPropagateResult::NO_PRUNING_POSSIBLE;
+	}
+	auto &data = input.bind_data->Cast<PrefixRangeFunctionData>();
+	if (!data.filter || !data.filter->IsInitialized()) {
+		return FilterPropagateResult::NO_PRUNING_POSSIBLE;
+	}
+	if (input.stats.GetStatsType() != StatisticsType::NUMERIC_STATS) {
+		return FilterPropagateResult::NO_PRUNING_POSSIBLE;
+	}
+	if (!NumericStats::HasMinMax(input.stats)) {
+		return FilterPropagateResult::NO_PRUNING_POSSIBLE;
+	}
+
+	const auto min = NumericStats::Min(input.stats);
+	const auto max = NumericStats::Max(input.stats);
+	if (min > max) {
+		return FilterPropagateResult::NO_PRUNING_POSSIBLE;
+	}
+
+	return data.filter->LookupRange(min, max);
+}
+
+ScalarFunction TableFilterPrefixRangeFun::GetFunction() {
+	return PrefixRangeScalarFun::GetFunction(LogicalType::ANY);
+}
+
+} // namespace duckdb
